Add More Pipe Flags to speed up compile
A common flag is -pipe . This flag actually has no effect on the generated code, but it makes the compilation process faster. It tells the compiler to use pipes instead of temporary files during the different stages of compilation, which uses more memory. On systems with low memory, GCC might get killed. In that case, do not use this flag.

Signed-off-by: Chet Kener <Cl3Kener@gmail.com>